GKarman/hw9 Добавил тип float

+++ hw9/SQL/ddl.sql
@@ -28,7 +28,7 @@ CREATE TABLE movies_attributes_values
     value_int            INTEGER,
     value_timestamp      TIMESTAMP,
     value_boolean        BOOLEAN,
-    value_float          NUMERIC(15, 4),
+    value_float          FLOAT,
 
     FOREIGN KEY (movie_id) REFERENCES movies (id),
     FOREIGN KEY (movies_attributes_id) REFERENCES movies_attributes (id)
